requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.31.1

# --- Database (PostgreSQL + ORM) ---
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2   # (opcional, para migraciones)

# --- Configuración / entorno ---
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.4.0  # (opcional, si luego quieres manejar .env más limpio)

# --- Google Cloud SDK / Vertex AI ---
google-cloud-aiplatform==1.72.0
google-cloud-storage==2.18.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1

# --- Utilidades ---
requests==2.32.3
httpx==0.27.2
